# SchoolAPI/Final_Project_Swagger_File.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "School API",
		"version": "v1"
	},
	"paths": {
...
		"/api/v1/Courses/{id}": {
			"get": {
				"tags": [
					"Courses"
				],
				"operationId": "GetACourse",
				"parameters": [
					{
						"name": "coursesectionID",
						"in": "query",
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			},
			"delete": {
				"tags": [
					"Courses"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			},
			"put": {
				"tags": [
					"Courses"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json-patch+json": {
							"schema": {
								"$ref": "#/components/schemas/CourseForUpdateDto"
							}
						},
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CourseForUpdateDto"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/CourseForUpdateDto"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/CourseForUpdateDto"
							}
						},
						"application/xml": {
							"schema": {
								"$ref": "#/components/schemas/CourseForUpdateDto"
							}
						},
						"text/xml": {
							"schema": {
								"$ref": "#/components/schemas/CourseForUpdateDto"
							}
						},
						"application/*+xml": {
							"schema": {
								"$ref": "#/components/schemas/CourseForUpdateDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			},
			"patch": {
				"tags": [
					"Courses"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json-patch+json": {
							"schema": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Operation"
								},
								"nullable": true
							}
						},
						"application/json": {
							"schema": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Operation"
								},
								"nullable": true
							}
						},
						"text/json": {
							"schema": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Operation"
								},
								"nullable": true
							}
						},
						"application/*+json": {
							"schema": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Operation"
								},
								"nullable": true
							}
						},
						"application/xml": {
							"schema": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Operation"
								},
								"nullable": true
							}
						},
						"text/xml": {
							"schema": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Operation"
								},
								"nullable": true
							}
						},
						"application/*+xml": {
							"schema": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Operation"
								},
								"nullable": true
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			}
		},
...
		"/api/v1/SectionEnrollmentManage": {
			"get": {
				"tags": [
					"SectionEnrollmentManage"
				],
				"parameters": [
					{
						"name": "PageNumber",
						"in": "query",
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"name": "PageSize",
						"in": "query",
						"schema": {
							"type": "integer",
							"format": "int32"
						}
					},
					{
						"name": "OrderBy",
						"in": "query",
						"schema": {
							"type": "string",
							"nullable": true
						}
					},
					{
						"name": "Fields",
						"in": "query",
						"schema": {
							"type": "string",
							"nullable": true
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			},
			"post": {
				"tags": [
					"SectionEnrollmentManage"
				],
				"operationId": "SectionEnrollManageByID",
				"requestBody": {
					"content": {
						"application/json-patch+json": {
							"schema": {
								"$ref": "#/components/schemas/SectionEnrollmentManageDto"
							}
						},
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SectionEnrollmentManageDto"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/SectionEnrollmentManageDto"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/SectionEnrollmentManageDto"
							}
						},
						"application/xml": {
							"schema": {
								"$ref": "#/components/schemas/SectionEnrollmentManageDto"
							}
						},
						"text/xml": {
							"schema": {
								"$ref": "#/components/schemas/SectionEnrollmentManageDto"
							}
						},
						"application/*+xml": {
							"schema": {
								"$ref": "#/components/schemas/SectionEnrollmentManageDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			}
		},
		"/api/v1/SectionEnrollmentManage/{id}": {
			"get": {
				"tags": [
					"SectionEnrollmentManage"
				],
				"operationId": "GetASectionEnrollmentManage",
				"parameters": [
					{
						"name": "coursesectionID",
						"in": "query",
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			},
			"delete": {
				"tags": [
					"SectionEnrollmentManage"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			},
			"put": {
				"tags": [
					"SectionEnrollmentManage"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json-patch+json": {
							"schema": {
								"$ref": "#/components/schemas/SectionEnrollmentManageForUpdateDto"
							}
						},
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SectionEnrollmentManageForUpdateDto"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/SectionEnrollmentManageForUpdateDto"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/SectionEnrollmentManageForUpdateDto"
							}
						},
						"application/xml": {
							"schema": {
								"$ref": "#/components/schemas/SectionEnrollmentManageForUpdateDto"
							}
						},
						"text/xml": {
							"schema": {
								"$ref": "#/components/schemas/SectionEnrollmentManageForUpdateDto"
							}
						},
						"application/*+xml": {
							"schema": {
								"$ref": "#/components/schemas/SectionEnrollmentManageForUpdateDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			},
			"patch": {
				"tags": [
					"SectionEnrollmentManage"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json-patch+json": {
							"schema": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Operation"
								},
								"nullable": true
							}
						},
						"application/json": {
							"schema": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Operation"
								},
								"nullable": true
							}
						},
						"text/json": {
							"schema": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Operation"
								},
								"nullable": true
							}
						},
						"application/*+json": {
							"schema": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Operation"
								},
								"nullable": true
							}
						},
						"application/xml": {
							"schema": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Operation"
								},
								"nullable": true
							}
						},
						"text/xml": {
							"schema": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Operation"
								},
								"nullable": true
							}
						},
						"application/*+xml": {
							"schema": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Operation"
								},
								"nullable": true
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Success"
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"CourseForUpdateDto": {
				"required": [
					"courseName",
					"createdDate",
					"description",
					"updatedDate"
				],
				"type": "object",
				"properties": {
					"courseName": {
						"maxLength": 25,
						"minLength": 10,
						"type": "string"
					},
					"description": {
						"maxLength": 50,
						"minLength": 10,
						"type": "string"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"additionalProperties": false
			},
			"OperationType": {
				"enum": [
					0,
					1,
					2,
					3,
					4,
					5,
					6
				],
				"type": "integer",
				"format": "int32"
			},
			"Operation": {
				"type": "object",
				"properties": {
					"value": {
						"type": "object",
						"additionalProperties": false,
						"nullable": true
					},
					"operationType": {
						"$ref": "#/components/schemas/OperationType"
					},
					"path": {
						"type": "string",
						"nullable": true
					},
					"op": {
						"type": "string",
						"nullable": true
					},
					"from": {
						"type": "string",
						"nullable": true
					}
				},
				"additionalProperties": false
			},
...
			"SectionEnrollmentManageDto": {
				"required": [
					"startDate"
				],
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"startDate": {
						"maxLength": 60,
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					},
					"sectionID": {
						"type": "string",
						"nullable": true
					},
					"userID": {
						"type": "string",
						"nullable": true
					}
				},
				"additionalProperties": false
			},
			"SectionEnrollmentManageForUpdateDto": {
				"required": [
					"startDate"
				],
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"startDate": {
						"maxLength": 60,
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					},
					"sectionID": {
						"type": "string",
						"nullable": true
					},
					"userID": {
						"type": "string",
						"nullable": true
					}
				},
				"additionalProperties": false
			},
...
		}
	}
}